server/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Image processing dependencies
pdf2image==1.17.0
Pillow==10.1.0

# Google AI dependencies  
google-generativeai==0.3.2

# Future Google Places API
# googlemaps==4.10.0  # Uncomment when implementing Places API

# Development dependencies (optional)
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0